Revert "More Doctests errors cleared"
This reverts commit 4c3bb61a

+++ pandas/core/apply.py
@@ -1250,9 +1250,8 @@ def reconstruct_func(
 
     Examples
     --------
-    >>> reconstruct_func(None, **{"foo": ("col", "min")}) # doctest: +SKIP
-    (True, defaultdict(<class 'list'>, {'col': ['min']}),
-     ('foo',), array([0], dtype=int64))
+    >>> reconstruct_func(None, **{"foo": ("col", "min")})
+    (True, defaultdict(<class 'list'>, {'col': ['min']}), ('foo',), array([0]))
 
     >>> reconstruct_func("min")
     (False, 'min', None, None)
@@ -1329,9 +1328,8 @@ def normalize_keyword_aggregation(
 
     Examples
     --------
-    >>> normalize_keyword_aggregation({"output": ("input", "sum")}) # doctest: +SKIP
-    (defaultdict(<class 'list'>, {'input': ['sum']}),('output',),
-     array([0], dtype=int64))
+    >>> normalize_keyword_aggregation({"output": ("input", "sum")})
+    (defaultdict(<class 'list'>, {'input': ['sum']}), ('output',), array([0]))
     """
     from pandas.core.indexes.base import Index
